+++ NEWS.md
@@ -1,3 +1,5 @@
+# scales (development version)
+
 # scales 1.1.1
 
 * `breaks_width()` now handles `difftime`/`hms` objects (@bhogan-mitre, #244).